[update] ログイン画面にリダイレクトさせる場合、ログイン完了後、元のページにリダイレクトさせるように変更

+++ ShoppingSite/src/com/internousdev/ShoppingSite/action/AdminDeleteItemAction.java
@@ -6,6 +6,7 @@ import org.apache.struts2.interceptor.SessionAware;
 
 import com.internousdev.ShoppingSite.dao.ItemDAO;
 import com.internousdev.ShoppingSite.util.CheckAdmin;
+import com.internousdev.ShoppingSite.util.CheckLogin;
 import com.opensymphony.xwork2.ActionSupport;
 
 public class AdminDeleteItemAction extends ActionSupport implements SessionAware
@@ -15,6 +16,11 @@ public class AdminDeleteItemAction extends ActionSupport implements SessionAware
 
 	public String execute()
 	{
+		if(!CheckLogin.IsLogin(session))
+		{
+			session.put("LoginedRedirectAction", "ItemListAction");
+			return "needLogin";
+		}
 		if(!CheckAdmin.IsAdmin(session)) return "notAdmin";
 
 		ItemDAO.DeleteItem(id);
